Current firmware (includes step status)

+++ firmware/src/console.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 
 // Returns number of bytes processed
-int_fast8_t console_receive(uint8_t *buf, uint_fast8_t buf_len);
+int_fast16_t console_receive(uint8_t *buf, uint_fast16_t buf_len);
 
-void console_send(const uint8_t *buf, uint_fast8_t buf_len);
+void console_send(const uint8_t *buf, uint_fast16_t buf_len);
 void console_send_str(char *str);
 void console_send_uint8(uint32_t val);
 void console_send_uint32(uint32_t val);
